Order Exception 추가 및 디렉토리 수정
