+++ src/main/java/shop/dalda/exception/template/OrderNotFoundException.java
@@ -1,14 +1,15 @@
-package shop.dalda.exception;
+package shop.dalda.exception.template;
+
 
 import org.springframework.http.HttpStatus;
 
-public class UserNotFoundException extends RuntimeException {
+public class OrderNotFoundException extends RuntimeException {
 
-    private final String ERROR_CODE = "U001";
+    private final String ERROR_CODE = "T002";
     private final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
-    private static final String MESSAGE = "사용자를 찾을 수 없습니다.";
+    private static final String MESSAGE = "템플릿을 찾을 수 없습니다.";
 
-    public UserNotFoundException() {
+    public OrderNotFoundException() {
         super(MESSAGE);
     }
 
@@ -20,3 +21,4 @@ public class UserNotFoundException extends RuntimeException {
         return HTTP_STATUS;
     }
 }
+
