+++ src/main/java/shop/dalda/exception/template/TemplateNotFoundException.java
@@ -1,14 +1,15 @@
-package shop.dalda.exception;
+package shop.dalda.exception.template;
+
 
 import org.springframework.http.HttpStatus;
 
-public class TemplateInvalidException extends RuntimeException {
+public class TemplateNotFoundException extends RuntimeException {
 
-    private final String ERROR_CODE = "T001";
+    private final String ERROR_CODE = "T002";
     private final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
-    private static final String MESSAGE = "템플릿의 형식이 올바르지 않습니다.";
+    private static final String MESSAGE = "템플릿을 찾을 수 없습니다.";
 
-    public TemplateInvalidException() {
+    public TemplateNotFoundException() {
         super(MESSAGE);
     }
 
@@ -20,3 +21,4 @@ public class TemplateInvalidException extends RuntimeException {
         return HTTP_STATUS;
     }
 }
+
